+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Muthukrishnan 2025 clazakizumab ([doi:10.1111/cts.70381](https://doi.org/10.1111/cts.70381)) -- adults with end-stage kidney disease receiving maintenance dialysis.
+- Add Jia 2025 TAK-071 ([doi:10.1002/cpdd.1579](https://doi.org/10.1002/cpdd.1579)) -- healthy adults and participants with Parkinson disease and cognitive impairment.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.